requirements.txt
fastapi
uvicorn
httpx
python-dotenv
aiogram
asyncpg==0.29.0

